Extracts arguments of all functions in a package into a dataframe

Extracts arguments of all functions in a package into a dataframe

rev_signature(package)

Arguments

package
to be inspected

Value

A dataframe with the following columns:

f_name: function name

f_args: function arguments

Details

The functions takes the name of an installed package

Examples

rev_signature(package = "graphics")

#> # A tibble: 125 x 2
#>    f_name          f_args                                                      
#>    <chr>           <chr>                                                       
#>  1 .filled.contour ".filled.contour (x, y, z, levels, col) "                   
#>  2 .SSassign       ".SSassign (x, value) "                                     
#>  3 .SSexists       ".SSexists (x) "                                            
#>  4 .SSget          ".SSget (x) "                                               
#>  5 abline          "abline (a = NULL, b = NULL, h = NULL, v = NULL, reg = NULL~
#>  6 arrows          "arrows (x0, y0, x1 = x0, y1 = y0, length = 0.25, angle = 3~
#>  7 assocplot       "assocplot (x, col = c(\"black\", \"red\"), space = 0.3, ma~
#>  8 axis            "axis (side, at = NULL, labels = TRUE, tick = TRUE, line = ~
#>  9 Axis            "Axis (x = NULL, at = NULL, ..., side, labels = NULL) "     
#> 10 axis.Date       "axis.Date (side, x, at, format, labels = TRUE, ...) "      
#> # ... with 115 more rows